Refactored and Renamed
