+++ CS320_Group_Project_Chess/main/edu/ycp/cs320/Group_Project_Chees/main/Main.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.lab02a_dmoore29.main;
+package edu.ycp.cs320.Group_Project_Chees.main;
 
 import java.io.File;
