+++ CS320_Group_Project_Chess/src/edu/ycp/cs320/Group_Project_Chees/controller/GuessingGameController.java
@@ -1,6 +1,6 @@
-package edu.ycp.cs320.lab02a_dmoore29.controller;
+package edu.ycp.cs320.Group_Project_Chees.controller;
 
-import edu.ycp.cs320.lab02a_dmoore29.model.GuessingGame;
+import edu.ycp.cs320.Group_Project_Chees.model.GuessingGame;
 
 /**
  * Controller for the guessing game.
