+++ CS320_Group_Project_Chess/src/edu/ycp/cs320/Group_Project_Chees/servlet/IndexServlet.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.lab02a_dmoore29.servlet;
+package edu.ycp.cs320.Group_Project_Chees.servlet;
 
 import java.io.IOException;
