+++ CS320_Group_Project_Chess/src/edu/ycp/cs320/Group_Project_Chees/servlet/MultiplyNumbersServlet.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.lab02a_dmoore29.servlet;
+package edu.ycp.cs320.Group_Project_Chees.servlet;
 
 import java.io.IOException;
 
@@ -7,8 +7,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.lab02a_dmoore29.controller.NumbersController;
-import edu.ycp.cs320.lab02a_dmoore29.model.Numbers;
+import edu.ycp.cs320.Group_Project_Chees.controller.NumbersController;
+import edu.ycp.cs320.Group_Project_Chees.model.Numbers;
 
 public class MultiplyNumbersServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
