+++ CS320_Group_Project_Chess/src/edu/ycp/cs320/Group_Project_Chees/servlet/ajax/AddNumbersAjaxServlet.java
@@ -1,4 +1,4 @@
-package edu.ycp.cs320.lab02a_dmoore29.servlet.ajax;
+package edu.ycp.cs320.Group_Project_Chees.servlet.ajax;
 
 import java.io.IOException;
 
@@ -7,10 +7,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import edu.ycp.cs320.lab02a_dmoore29.controller.NumbersController;
-import edu.ycp.cs320.lab02a_dmoore29.model.Numbers;
+import edu.ycp.cs320.Group_Project_Chees.controller.NumbersController;
+import edu.ycp.cs320.Group_Project_Chees.model.Numbers;
 
-public class MultiplyNumbersAjaxServlet extends HttpServlet {
+public class AddNumbersAjaxServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
 	@Override
@@ -29,9 +29,10 @@ public class MultiplyNumbersAjaxServlet extends HttpServlet {
 		// Get parameters
 		Double first = getDouble(req, "first");
 		Double second = getDouble(req, "second");
+		Double third = getDouble(req, "third");
 		
 		// Check whether parameters are valid
-		if (first == null || second == null) {
+		if (first == null || second == null || third == null) {
 			badRequest("Bad parameters", resp);
 			return;
 		}
@@ -39,15 +40,17 @@ public class MultiplyNumbersAjaxServlet extends HttpServlet {
 		
 		model.setFirst(first);
 		model.setSecond(second);
-		
+		model.setThird(third);
+				
 		// Use a controller to process the request
 		NumbersController controller = new NumbersController();
-		controller.multiply(first, second);
+		controller.add(first, second, third);
 		
 		// Send back a response
 		resp.setContentType("application/json");
 		resp.getWriter().println("{ \"first\": " + model.getFirst() +
 				", \"second\": " + model.getSecond() +
+				", \"third\": " + model.getThird() +
 				", \"result\": " + model.getResult() + "}" );	}
 
 	private Double getDouble(HttpServletRequest req, String name) {
